increase job channel size
When the GC goroutine kicks in an potentially wants to close N writers, it keeps a lock on the writers map. This means that the worker goroutine cannot close any writer until the GC is done. If N is >= size(job queue), the GC routine cannot write all the required jobs and release the lock, but at the same time the worker routine cannot do its job and decrease the job queue.